[storage/innobase] PS-7949: Stop buf_lru_manager_thread during calls to buf_pool_invalidate_instance
https://jira.percona.com/browse/PS-7949

---

*Problem*:

For buffer pool invalidation to proceed we must ensure there is no
LRU write activity happening.

PS does not implement this guarantee as PS implements a multi-threaded
LRU Manager unlike MySQL that uses a single thread.

This error manifests when running the test `clone.remote_ddl_basic`. It
is a not deterministic error due to the interleaving of execution of
several threads.

*Solution*:

`buf_pool_invalidate_instance` reset the  the event
`buf_pool_t->run_lru` and `buf_lru_manager_thread` waits for the event
`buf_pool_t->run_lru` to be set.