PS-8433: Reduce memory overhead from cached triggers by removing fixed size buffer for error.
https://jira.percona.com/browse/PS-8433

For tables with triggers, each TABLE instance in Table Cache has
one Table_trigger_dispatcher sub-object and one or more Trigger
sub-object associated with it.
Both these classes contain as members 512-byte buffer to store error
message which occurred during trigger parsing. Since these buffers
are not used in 99% of cases this looks like pure waste of memory.

This patch improves the situation by replacing these always-allocated
fixed-size buffers with variable-size buffers which are allocated
on TABLE's memory root on demand, only when the table has a trigger
with a parse error.