[plugin/clone] PS-8188 : Make the clone_plugin to not force plugins to match on recipient/donor (#5032)
PS-8188 : Make the clone_plugin to not force plugins to match on recipient/donor

https://jira.percona.com/browse/PS-8188

Problem:
--------
Clone enforces donor and recipient have same plugins installed.
For example, if a donor has an extra authentication plugin or a data masking plugin
installed but not installed on recipient, it is not possible to clone.
It will fail with the below error.

'Clone Donor plugin <plugin_name> is not active in Recipient.'

Fix:
----
Introduce new clone option 'clone_exclude_plugins_match' to take list of plugins that
are not installed on recipient.

This option is used on recipient server. The following plugins are disallowed. They must
be installed on recipient.

"daemon_keyring_proxy_plugin",
"binlog",
"performance_schema",
"memory",
"innodb",
"keyring_file",
"keyring_vault"

Co-authored-by: Satya Bodapati <satya.bodapati@percona.com>